# WIP: Avoid copying LogicalPlans / Exprs during OptimizerPasses