--- a/datafusion/expr/src/logical_plan/plan.rs
+++ b/datafusion/expr/src/logical_plan/plan.rs
@@ -361,10 +361,12 @@ impl LogicalPlan {
         }
     }
 
-    /// takes all inputs of this plan, unwrapping them if they are
-    /// not shared
-    pub fn take_inputs(&self) -> Vec<LogicalPlan> {
-        todo!()
+    /// applies the closure `f` to each input of this node, replacing the existing inputs
+    /// with the result of the closure.
+    pub fn rewrite_inputs<F>(mut self, f: F) -> Result<Self>
+    where F: FnMut (LogicalPlan) -> Result<LogicalPlan>
+    {
+        todo!();
     }
 
     /// returns all inputs of this `LogicalPlan` node. Does not
@@ -521,15 +523,6 @@ impl LogicalPlan {
     #[deprecated(since = "35.0.0", note = "please use `with_new_exprs` instead")]
     pub fn with_new_inputs(&self, inputs: &[LogicalPlan]) -> Result<LogicalPlan> {
         self.with_new_exprs(self.expressions(), inputs.to_vec())
-    }
-
-    /// returns a new LogicalPlan with the new inputs (potentially rewritten)
-    ///
-    pub fn with_new_inputs2(
-        mut self,
-        new_inputs: Vec<LogicalPlan>,
-    ) -> Result<LogicalPlan> {
-        todo!()
     }
 
     /// Returns a new `LogicalPlan` based on `self` with inputs and