## Which issue does this PR close?

Related to https://github.com/apache/arrow-datafusion/issues/9637

## Rationale for this change

@jayzhan211  @mustafasrepo  and I have been discussing how to speed up planning by avoiding copying Exprs  / LogicalPlans


## What changes are included in this PR?

This PR is a prototype of a suggestion here https://github.com/apache/arrow-datafusion/issues/9637#issuecomment-2002396168 to see if I can get a significant boost in Expr simplify

The theory is that if I can rerite the optimizer to take ownership of the plans they can be rewritten in place

If this works out, I'll turn this into an actual PR for review. 

Update: this approach improves planning speed by 25%

## Are these changes tested?

NA

Timings (main):
```
cargo bench   --bench sql_planner -- physical_plan_tpch_all

Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.4s, or reduce sample count to 60.
physical_plan_tpch_all  time:   [70.511 ms 71.102 ms 71.828 ms]
                        change: [-1.3186% +2.1648% +4.6865%] (p = 0.17 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) low mild
  1 (1.00%) high mild
  4 (4.00%) high severe
```

This branch
```
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.4s, or reduce sample count to 90.
physical_plan_tpch_all  time:   [54.371 ms 54.512 ms 54.675 ms]
                        change: [-24.152% -23.333% -22.648%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->