@alamb 2024-03-24
Continuing in https://github.com/apache/arrow-datafusion/pull/9780, so closing this one